requirements.txt
avro-python3==1.8.1
beautifulsoup4==4.4.1
elasticsearch==2.3.0
elasticsearch-dsl==2.0.0
mmh3==2.3.1
python-dateutil==2.5.2
requests==2.9.1
six==1.10.0
urllib3==1.14
